+++ app/src/lib/components/game/GameRoundWinner.svelte
@@ -100,9 +100,9 @@
             class="flex items-center gap-2 px-1 text-xl text-gray-200 md:py-1 [&:not(:last-child)]:border-b border-fuchsia-100/50"
           >
             <span class=""
-              >{posLast > posTotal
+              >{posLast > pos
                 ? '🔼'
-                : posLast === posTotal
+                : posLast === pos
                   ? '➖'
                   : '🔽'}</span
             >
