+++ app/src/lib/icons/IconMinimize.svelte
@@ -1,0 +1,16 @@
+<svg
+  aria-label="Pantalla completa"
+  xmlns="http://www.w3.org/2000/svg"
+  width="24"
+  height="24"
+  viewBox="0 0 24 24"
+  fill="none"
+  stroke="currentColor"
+  stroke-width="2"
+  stroke-linecap="round"
+  stroke-linejoin="round"
+  class="lucide lucide-minimize"
+  ><path d="M8 3v3a2 2 0 0 1-2 2H3" /><path d="M21 8h-3a2 2 0 0 1-2-2V3" /><path
+    d="M3 16h3a2 2 0 0 1 2 2v3"
+  /><path d="M16 21v-3a2 2 0 0 1 2-2h3" /></svg
+>
