+++ app/src/lib/icons/IconVolumex.svelte
@@ -1,0 +1,19 @@
+<svg
+  aria-label="Desmutear"
+  xmlns="http://www.w3.org/2000/svg"
+  width="24"
+  height="24"
+  viewBox="0 0 24 24"
+  fill="none"
+  stroke="currentColor"
+  stroke-width="2"
+  stroke-linecap="round"
+  stroke-linejoin="round"
+  class="lucide lucide-volume-x"
+  ><polygon points="11 5 6 9 2 9 2 15 6 15 11 19 11 5" /><line
+    x1="22"
+    x2="16"
+    y1="9"
+    y2="15"
+  /><line x1="16" x2="22" y1="9" y2="15" /></svg
+>
